Remove duplication, uneeded upsert
